requirements.txt
paho-mqtt
zeroconf
ipaddress
requests
phue
py-applescript
